backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
Werkzeug==3.0.2
pymongo==4.6.3